app/backend-v1/admin/requirements.txt
boto3==1.20.42
uvicorn
fastapi
python-dotenv
python-multipart